Fixed a bug where `onClose` event was not notified on closing of the WebSocket connection
We were cleaning up the socket right after the `close` call. Cleaning up
the socket means removing the actual event listeners.

Hence, by clean up at this point, the `onclose` event was not be handled
and the consumer of the library was not receiving its own `onClose` event.

Instead, let's do the clean up once we receive the socket `onclose` event.