Fixed a bug where `onClose` event was not notified anymore
When an error occurs, the WebSocket protocol specifies that an `onerror`
event is being emitted followed by an `onclose` event. However, we were
actually cleaning up the socket in the `onerror` handler leading the
`onclose` event to be swalloed altogether and not reported to the
consumer of the library anymore

This was fixed by reverting the offending code and let the `onclose`
handler always clean the connection instead.