Fixed up `scripts/gh-push.js` not updating correctly
By using a direct url, the internal git database of the `origin` remote
is not updated on the push leading to the next status to think we are
ahead of remote tip since the internal database does not "see" the commits
we push through the url directly.

The idea of using the url was that later, the CI could have a
`GITHUB_API_TOKEN` and would be able to push the data to the repository.
A workaround we could do would be to use the direct url only when the
`GITHUB_API_TOKEN` is set.

Reverted to use `origin` for now until we have the need for the CI.