Added `autoDisconnectSocket` option (on standard WebSocket streaming client)
This adds the possibility to prevent automatic disconnection of socket when no more
stream are active in the client. This was already available for GraphQL over WebSocket
but not for standard WebSocket stream.

Helps on cases where the consumer know that it will most probably start a new stream
in the foreseeable future.