Fixed race condition when starting to start new streams in some conditions
When the socket is currently disconnecting, the logic flow is split in two parts. First
we disconnect the socket in the first operation and second, we wait for the socket
`onclose` notification to resolve the disconnection promise.

Between the two, if a new connection attempt was tried, there was chances the `onclose`
notification of the second phase of disconnection would clear some state initialized by
the new connection trying to start.

This state clearing was causing some of the new streams waiting for the connection to
complete to never resolve correctly.

Now, if a disconnection is in progress when trying to connect a new socket, we first wait
for the second phase of the disconnection to complete before continuing. This way, no
in-between state is cleared.

This should not have any other impact, `onclose` notification are call really rapidly
so there should be no problem around this related to timing.